+++ ChangeLog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Sparse Voxel Convolution
+- Clebsch-Gordan coefficients are computed via a change of basis from the complex to real basis.
 - `o3`, `nn` and `io` are accessible through `e3nn`. For instance `e3nn.o3.rand_axis_angle`.
 
 ### Changed